+++ py/desisurvey/scripts/surveymovie.py
@@ -7,6 +7,7 @@ command search path.
 The optional matplotlib python package must be installed to use this script.
 
 The external program ffmpeg must be installed to use this script.
+At nersc, try ``module add ffmpeg``.
 """
 from __future__ import print_function, division, absolute_import
 
@@ -59,8 +60,16 @@ def parse(options=None):
         '--expid', type=int, default=None, metavar='ID',
         help='index of single exposure to display')
     parser.add_argument(
+        '--nightly', action='store_true',
+        help='output one summary frame per night')
+    parser.add_argument(
+        '--scores', action='store_true', help='display scheduler scores')
+    parser.add_argument(
         '--save', type=str, default='surveymovie', metavar='NAME',
         help='base name (without extension) of output file to write')
+    parser.add_argument(
+        '--fps', type=float, default=10., metavar='FPS',
+        help='frames per second to render')
     parser.add_argument(
         '--label', type=str, default='DESI', metavar='TEXT',
         help='label to display on each frame')
@@ -75,6 +84,10 @@ def parse(options=None):
         args = parser.parse_args()
     else:
         args = parser.parse_args(options)
+
+    if args.nightly and args.scores:
+        log.warn('Cannot display scores in nightly summary.')
+        args.scores = False
 
     # Validate start/stop date args and covert to datetime objects.
     # Unspecified values are taken from our config.
@@ -108,20 +121,23 @@ def wrap(angle, offset=-60):
 class Animator(object):
     """Manage animation of survey progress.
     """
-    def __init__(self, ephem, progress, start, stop, label):
-        """
-        """
+    def __init__(self, ephem, progress, start, stop, label, show_scores):
+        self.log = desiutil.log.get_logger()
         self.ephem = ephem
         self.progress = progress
         self.label = label
+        self.show_scores = show_scores
         self.config = desisurvey.config.Configuration()
         tiles = progress._table
         self.ra = wrap(tiles['ra'])
         self.dec = tiles['dec']
         self.passnum = tiles['pass']
         self.tileid = tiles['tileid']
-        self.tiles_per_pass = np.count_nonzero(
-            self.passnum == np.arange(8).reshape(8, 1), axis=1)
+        npass = np.max(self.passnum) + 1
+        self.tiles_per_pass = np.zeros(npass, int)
+        for passnum in np.unique(self.passnum):
+            self.tiles_per_pass[passnum] = np.count_nonzero(
+                self.passnum == passnum)
         self.prognames = ['DARK', 'DARK', 'DARK', 'DARK', 'GRAY',
                           'BRIGHT', 'BRIGHT', 'BRIGHT']
         self.psels = [
@@ -136,6 +152,7 @@ class Animator(object):
             start, stop, tile_fields='tileid,index,ra,dec,pass',
             exp_fields='expid,mjd,night,exptime,snr2cum,seeing,transparency')
         self.num_exp = len(self.exposures)
+        self.num_nights = len(np.unique(self.exposures['night']))
 
         # Calculate each exposure's LST window.
         exp_midpt = astropy.time.Time(
@@ -148,8 +165,8 @@ class Animator(object):
         self.lst[:, 0] = wrap(lst_midpt - 0.5 * lst_len)
         self.lst[:, 1] = wrap(lst_midpt + 0.5 * lst_len)
 
-    def init_figure(self, width=1920, height=1080, dpi=32):
-        """
+    def init_figure(self, nightly, width=1920, height=1080, dpi=32):
+        """Initialize matplot artists for drawing each frame.
         """
         self.dpi = float(dpi)
         # Initialize figure and axes.
@@ -171,7 +188,9 @@ class Animator(object):
         self.f_obj = [None] * navoids
         bgcolor = matplotlib.colors.to_rgba('lightblue')
         avoidcolor = matplotlib.colors.to_rgba('red')
+        self.defaultcolor = np.array([[1., 1., 1., 1.]])
         self.completecolor = np.array([0., 0.5, 0., 1.])
+        self.availcolor = np.array([1., 1., 1., 1.])
         self.unavailcolor = np.array([0.65, 0.65, 0.65, 1.])
         self.nowcolor = np.array([0., 0.7, 0., 1.])
         pcolors = desisurvey.plots.program_color
@@ -214,9 +233,8 @@ class Animator(object):
                 # Draw the tile outlines for this pass.
                 sel = (self.passnum == passnum)
                 ntiles = np.count_nonzero(sel)
-                fc = np.zeros((ntiles, 4))
-                fc[:, 1] = 1.0
-                fc[:, 3] = 0.5
+                fc = np.empty((ntiles, 4))
+                fc[:] = self.defaultcolor
                 s = np.full(ntiles, 90.)
                 self.scatters.append(ax.scatter(
                     self.ra[sel], self.dec[sel], s=s, facecolors=fc,
@@ -226,18 +244,20 @@ class Animator(object):
                 ec = np.zeros((navoids, 4))
                 ec[:] = avoidcolor
                 s = np.full(navoids, 500.)
-                s[0] = 2500.
+                s[self.moon_index] = 2500.
                 self.avoids.append(ax.scatter(
                     self.xy_avoid[:, 0], self.xy_avoid[:, 1], facecolors=fc,
                     edgecolors=ec, s=s, lw=2))
-                # Draw LST lines for the current exposure.
-                line1 = ax.axvline(0., lw=2, ls=':', color=self.nowcolor)
-                line2 = ax.axvline(0., lw=2, ls=':', color=self.nowcolor)
-                self.lstlines.append((line1, line2))
+                if not nightly:
+                    # Draw LST lines for the current exposure.
+                    line1 = ax.axvline(0., lw=2, ls=':', color=self.nowcolor)
+                    line2 = ax.axvline(0., lw=2, ls=':', color=self.nowcolor)
+                    self.lstlines.append((line1, line2))
                 passnum += 1
 
-        # Initialize scheduler score colormap.
-        self.scorecmap = matplotlib.cm.get_cmap('hot_r')
+        if self.show_scores:
+            # Initialize scheduler score colormap.
+            self.scorecmap = matplotlib.cm.get_cmap('hot_r')
 
         # Add axis above plot to display programs during the night.
         paxes = plt.axes([0, 0.97, 0.66667, 0.03], facecolor='y')
@@ -255,8 +275,9 @@ class Animator(object):
             self.pdata.reshape(1, -1), interpolation='none', aspect='auto',
             extent=(edges[0], edges[-1], 0., 1.),
             vmin=-0.5, vmax=3.5, cmap=pcmap)
-        self.pline1 = paxes.axvline(0., lw=4, ls='-', color=self.nowcolor)
-        self.pline2 = paxes.axvline(0., lw=4, ls='-', color=self.nowcolor)
+        if not nightly:
+            self.pline1 = paxes.axvline(0., lw=4, ls='-', color=self.nowcolor)
+            self.pline2 = paxes.axvline(0., lw=4, ls='-', color=self.nowcolor)
 
         # Add text label in the top-right corner.
         self.text = plt.annotate(
@@ -266,39 +287,52 @@ class Animator(object):
 
         # List all animated artists.
         self.artists = (
-            self.scatters + self.avoids + self.labels + self.iplots + [
-            self.programs, self.pline1, self.pline2, self.text])
-        for l1, l2 in self.lstlines:
-            self.artists += [l1, l2]
+            self.scatters + self.avoids + self.labels + self.iplots +
+            [self.programs, self.text])
+        if not nightly:
+            self.artists += [self.pline1, self.pline2]
+            for l1, l2 in self.lstlines:
+                self.artists += [l1, l2]
 
         # Initialize internal tracking vars.
         self.last_date = None
         self.scores = None
-        self.idx0 = None
+        self.iexp0 = None
         self.status = None
 
-    def init_date(self, date, night):
-        """
+    def init_date(self, date, ephem):
+        """Initialize before drawing frames for a new night.
+
+        Parameters
+        ----------
+        date : datetime.date
+            Date on which this night's observing starts.
+        night : astropy.table.Column
+            Ephemerides data for this night.
         """
         # Update the observing program for this night.
-        dark, gray, bright = self.ephem.get_program(night['noon'] + self.dmjd)
+        dark, gray, bright = self.ephem.get_program(ephem['noon'] + self.dmjd)
         self.pdata[:] = 0
         self.pdata[dark] = 1
         self.pdata[gray] = 2
         self.pdata[bright] = 3
         self.programs.set_data(self.pdata.reshape(1, -1))
-        # Load new scheduler scores for this night.
-        scores_name = self.config.get_path('scores_{0}.fits'.format(date))
-        if os.path.exists(scores_name):
-            hdus = astropy.io.fits.open(scores_name, memmap=False)
-            self.scores = hdus[0].data
-            hdus.close()
-            # Save index of first exposure on this date.
-            self.idx0 = np.argmax(self.exposures['night'] == str(date))
+        if self.show_scores:
+            # Load new scheduler scores for this night.
+            scores_name = self.config.get_path('scores_{0}.fits'.format(date))
+            if os.path.exists(scores_name):
+                hdus = astropy.io.fits.open(scores_name, memmap=False)
+                self.scores = hdus[0].data
+                hdus.close()
+                # Save index of first exposure on this date.
+                noon = desisurvey.utils.local_noon_on_date(date)
+                self.iexp0 = np.argmax(self.exposures['mjd'] > noon.mjd)
+            else:
+                self.warn('Missing scores file: {0}.'.format(scores_name))
         # Get interpolator for moon, planet positions during this night.
         for i, name in enumerate(self.avoid_names):
             self.f_obj[i] = desisurvey.ephemerides.get_object_interpolator(
-                night, name)
+                ephem, name)
         if self.last_date is not None:
             week_num = int(np.floor((date - self.start_date).days / 7.))
             # Update progress graphs for each program.
@@ -308,57 +342,80 @@ class Animator(object):
                 yprog = iplot.get_ydata()
                 yprog[week_num] = 1.0 * ndone / nprog
                 iplot.set_ydata(yprog)
-        # Lookup tonight's plan.
-        plan_name = self.config.get_path('plan_{0}.fits'.format(date))
-        plan = astropy.table.Table.read(plan_name)
-        self.available = plan['available']
-        self.priority = plan['priority']
+        # Lookup which tiles are available and planned for tonight.
+        day_number = desisurvey.utils.day_number(date)
+        avail = self.progress._table['available']
+        self.available = (avail >= 0) & (avail <= day_number)
+        planned = self.progress._table['planned']
+        self.planned = (planned >= 0) & (planned <= day_number)
         self.last_date = date
 
-    def draw_exposure(self, idx, last_date=None, scores=None, idx0=0):
+    def draw_exposure(self, iexp, nightly):
+        """Draw the frame for a single exposure.
+
+        Calls :meth:`init_date` if this is the first exposure of the night
+        that we have seen.
+
+        Parameters
+        ----------
+        iexp : int
+            Index of the exposure to draw.
+
+        Returns
+        -------
+        bool
+            True if a new frame was drawn.
         """
-        """
-        info = self.exposures[idx]
+        info = self.exposures[iexp]
         mjd = info['mjd']
         date = desisurvey.utils.get_date(mjd)
-        assert str(date) == info['night']
+        assert date == desisurvey.utils.get_date(info['night'])
         night = self.ephem.get_night(date)
         # Initialize status if necessary.
-        if self.status is None:
+        if (self.status is None) or nightly:
             snapshot = self.progress.copy_range(mjd_max=mjd)
             self.status = np.array(snapshot._table['status'])
         if date != self.last_date:
             # Initialize for this night.
             self.init_date(date, night)
+        elif nightly:
+            return False
         # Update the status for the current exposure.
         complete = info['snr2cum'] >= self.config.min_snr2_fraction()
         self.status[info['index']] = 2 if complete else 1
         # Update the top-right label.
-        self.text.set_text(
-            '{0} {1} #{2:06d} ({3:.1f}",{4:.2f})'
-            .format(self.label, date, info['expid'], info['seeing'],
-                    info['transparency']))
-        # Update current time in program.
-        dt1 = mjd - night['noon']
-        dt2 = dt1 + info['exptime'] / 86400.
-        self.pline1.set_xdata([dt1, dt1])
-        self.pline2.set_xdata([dt2, dt2])
-        # Update scores display for this exposure.
-        score = self.scores[idx - self.idx0]
-        max_score = np.max(score)
+        label = '{} {} #{:06d}'.format(self.label, date, info['expid'])
+        if not nightly:
+            label += ' ({:.1f}",{:.2f})'.format(
+                info['seeing'], info['transparency'])
+        self.text.set_text(label)
+        if not nightly:
+            # Update current time in program.
+            dt1 = mjd - night['noon']
+            dt2 = dt1 + info['exptime'] / 86400.
+            self.pline1.set_xdata([dt1, dt1])
+            self.pline2.set_xdata([dt2, dt2])
+        if self.show_scores:
+            # Update scores display for this exposure.
+            score = self.scores[iexp - self.iexp0]
+            max_score = np.max(score)
         for passnum, scatter in enumerate(self.scatters):
             sel = (self.passnum == passnum)
-            fc = self.scorecmap(score[sel] / max_score)
             done = self.status[sel] == 2
-            inplan = self.priority[sel] > 0
             avail = self.available[sel]
+            inplan = self.planned[sel]
+            if self.show_scores:
+                fc = self.scorecmap(score[sel] / max_score)
+            else:
+                fc = scatter.get_facecolors()
+                fc[avail] = self.availcolor
             sizes = scatter.get_sizes()
             sizes[~inplan] = 20.
             sizes[~done & inplan] = 90.
             sizes[done] = 30.
             fc[done] = self.completecolor
             fc[~avail] = self.unavailcolor
-            if info['pass'] == passnum:
+            if not nightly and (info['pass'] == passnum):
                 # Highlight the tile being observed now.
                 jdx = np.where(self.tileid[sel] == info['tileid'])[0][0]
                 fc[jdx] = self.nowcolor
@@ -369,14 +426,15 @@ class Animator(object):
                    self.tiles_per_pass[passnum])
             self.labels[passnum].set_text('{0}-{1} {2:5.1f}%'.format(
                 self.prognames[passnum], passnum, pct))
-        # Update LST lines.
-        x1, x2 = self.lst[idx]
-        for passnum, (line1, line2) in enumerate(self.lstlines):
-            ls = '-' if info['pass'] == passnum else '--'
-            line1.set_linestyle(ls)
-            line2.set_linestyle(ls)
-            line1.set_xdata([x1, x1])
-            line2.set_xdata([x2, x2])
+        if not nightly:
+            # Update LST lines.
+            x1, x2 = self.lst[iexp]
+            for passnum, (line1, line2) in enumerate(self.lstlines):
+                ls = '-' if info['pass'] == passnum else '--'
+                line1.set_linestyle(ls)
+                line2.set_linestyle(ls)
+                line1.set_xdata([x1, x1])
+                line2.set_xdata([x2, x2])
         # Fill the moon with a shade of gray corresponding to its illuminated
         # fraction during this exposure.
         moon_frac = self.ephem.get_moon_illuminated_fraction(mjd)
@@ -389,6 +447,7 @@ class Animator(object):
             scatter.set_offsets(self.xy_avoid)
             scatter.get_facecolors()[self.moon_index] = [
                 moon_frac, moon_frac, moon_frac, 1.]
+        return True
 
 
 def main(args):
@@ -418,10 +477,12 @@ def main(args):
     progress = desisurvey.progress.Progress(args.progress)
 
     # Initialize.
-    animator = Animator(ephem, progress, args.start, args.stop, args.label)
-    log.info('Found {0} exposures from {1} to {2}.'
-             .format(animator.num_exp, args.start, args.stop))
-    animator.init_figure()
+    animator = Animator(
+        ephem, progress, args.start, args.stop, args.label, args.scores)
+    log.info('Found {0} exposures from {1} to {2} ({3} nights).'
+             .format(animator.num_exp, args.start, args.stop,
+                     animator.num_nights))
+    animator.init_figure(args.nightly)
 
     if args.expid is not None:
         expid = animator.exposures['expid']
@@ -429,24 +490,31 @@ def main(args):
         if (args.expid < expid[0]) or (args.expid > expid[-1]):
             raise RuntimeError('Requested exposure ID {0} not available.'
                                .format(args.expid))
-        animator.draw_exposure(args.expid - expid[0])
+        animator.draw_exposure(args.expid - expid[0], args.nightly)
         save_name = args.save + '.png'
         plt.savefig(save_name)
         log.info('Saved {0}.'.format(save_name))
     else:
+        nframes = animator.num_nights if args.nightly else animator.num_exp
+        iexp = [0]
         def init():
             return animator.artists
-        def update(idx):
-            if (idx + 1) % args.log_interval == 0:
+        def update(iframe):
+            if (iframe + 1) % args.log_interval == 0:
                 log.info('Drawing frame {0}/{1}.'
-                         .format(idx + 1, animator.num_exp))
-            animator.draw_exposure(idx)
+                         .format(iframe + 1, nframes))
+            if args.nightly:
+                while not animator.draw_exposure(iexp[0], nightly=True):
+                    iexp[0] += 1
+            else:
+                animator.draw_exposure(iexp=iframe, nightly=False)
             return animator.artists
+        log.info('Movie will be {:.1f} mins long at {:.1f} frames/sec.'
+                 .format(nframes / (60 * args.fps), args.fps))
         animation = matplotlib.animation.FuncAnimation(
-            animator.figure, update, init_func=init, interval=100,
-            blit=True, frames=animator.num_exp)
+            animator.figure, update, init_func=init, blit=True, frames=nframes)
         writer = matplotlib.animation.writers['ffmpeg'](
-            bitrate=2400, metadata=dict(artist='surveymovie'))
+            bitrate=2400, fps=args.fps, metadata=dict(artist='surveymovie'))
         save_name = args.save + '.mp4'
         animation.save(save_name, writer=writer, dpi=animator.dpi)
         log.info('Saved {0}.'.format(save_name))