+++ doc/changes.rst
@@ -5,7 +5,11 @@ desisurvey change log
 0.9.3 (unreleased)
 ------------------
 
-* No changes yet.
+* Fixes #18, #49, #54.
+* Improvements to surveymovie script.
+* Add progress columns to track fiber assignment and planning.
+* Add support for optional depth-first survey strategy.
+* Docs now auto-generated at http://desisurvey.readthedocs.io/en/latest/
 
 0.9.2 (2017-09-29)
 ------------------